refactor: 예외처리 폴더 생성 후 이동
